work around to fix publishing error
The components "vptstream" and "depthkit-player" both create threejs object3d in runtime. These objects don't need to be in the glb/gltf export as the Hubs client will also create them. For some reason the dynamically created object3d don't get cloned properly in the export process, the result is the process errors out.

This commit hacks around it by filtering for the nodeName property "NotForExport".  VPTStream and DepthKit set this when creating the object.